Trying with srsEPC

+++ scripts/ran/proxy_setup.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [ -f /local/repository/proxy-setup-complete ]; then
+    echo "Proxy setup already ran; not running again"
+    exit 0
+fi
+
 # Install dependencies
 sudo apt -y update
 sudo apt -y install libsctp-dev
